:+1: emotion を使った globalStyles の指定をするよう変更
