--- css/style.css
@@ -1,79 +1,0 @@
-@charset "UTF-8";
-
-html {
-  font-size: 62.5%;
-}
-
-body {
-  color: #333;
-  font-size: 1.2rem;
-  font-family: "Hiragino Kaku Gothic ProN", Meiryo, sans-serif;
-}
-
-*, *::before, *::after {
-  box-sizing: border-box;
-}
-a:link, a:visited, a:hover, a:active {
-  color: #ffa500;
-  text-decoration: none;
-}
-
-/* --- header ---------- */
-
-.header {
-  width: 100%;
-  padding: 28px 0 10px;
-  border-bottom: 3px double #589738;
-  box-shadow: 0 0 10px #e3e3e3;
-}
-.header .title {
-  width: 970px;
-  margin: 10px auto;
-  font-weight: normal;
-  letter-spacing: 1px;
-}
-
-/* --- main ---------- */
-
-.main {
-  width: 970px;
-  height: 300px;
-  margin: 30px auto 40px;
-}
-
-.main .title {
-  margin: 15px 0;
-  padding-left: 8px;
-  border-left: 5px solid #589738;
-  font-size: 1.4rem;
-  line-height: 1.2;
-}
-
-.main ul {
-  list-style-type: none;
-  margin: 15px;
-}
-
-.main ul > li {
-  margin-bottom: 15px;
-}
-
-.main ul > li a:hover {
-  text-decoration: underline;
-}
-
-
-/* --- footer ---------- */
-
-.footer {
-  width: 100%;
-  padding: 20px 0 30px;
-  background-color: #589738;
-  color: #fff;
-}
-
-.copyright {
-  margin-top: 30px;
-  text-align: center;
-  letter-spacing: 1px;
-}
