+++ package.json
@@ -12,9 +12,7 @@
     "start": "next start"
   },
   "dependencies": {
-    "@emotion/core": "^11.0.0",
-    "@emotion/styled": "^11.0.0",
-    "babel-plugin-emotion": "^11.0.0",
+    "@emotion/react": "^11.1.4",
     "next": "^10.0.5",
     "react": "^17.0.1",
     "react-dom": "^17.0.1"
