+++ pages/_app.tsx
@@ -1,6 +1,12 @@
-import '../src/styles.css'
+import { globalStyles } from '../styles/global-style'
+
 
 // This default export is required in a new `pages/_app.js` file.
 export default function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
+  return (
+    <>
+      {globalStyles}
+      <Component {...pageProps} />
+    </>
+  )
 }
