--- src/components/footer.tsx
@@ -1,21 +1,0 @@
-const Footer = () => (
-  <footer>
-    <p>Copyright (c) 2017-2020 fanjia38.github.io</p>
-
-    <style jsx>{`
-    footer {
-      width: 100%;
-      font-size: 0.8rem;
-      border-top: 3px double #589738;
-    }
-
-    p {
-      margin-top: 30px;
-      text-align: center;
-      letter-spacing: 1px;
-    }
-    `}</style>
-  </footer>
-);
-
-export default Footer;
