--- src/styles.css
@@ -1,28 +1,0 @@
-html,
-body {
-  padding: 0;
-  margin: 0;
-  font-family: 'Montserrat', sans-serif;
-  line-height: 1.6;
-  font-size: 18px;
-  color: #333;
-}
-
-*, *::before, *::after {
-  box-sizing: border-box;
-}
-
-a {
-  color: #0070f3;
-  text-decoration: none;
-}
-
-a:hover, a:visited, a:link, a:active {
-  text-decoration: none;
-  color: #ffa500;
-}
-
-img {
-  max-width: 100%;
-  display: block;
-}
